Backend/requirements.txt
flask>=3.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow==11.3.0
numpy<2.0.0
python-multipart>=0.0.9
pydantic>=2.6.0
werkzeug>=3.0.0
h5py>=3.9.0
opencv-python-headless>=4.9.0
tensorflow-cpu  # Add this if it's missing to ensure the model can load
flask-cors
